Updating Fields. Adding Components
Adding the Semantic Card thematic, with the CardHeader and CardBody,
Adding an info message for the Time Input.

Adding a Loading Spinner to erase once the user make an entry of a new
subject to the app.

Adding the Comments for the type of direction on the table,
Adding the table to a dev stage to implement the inputs.